pyproject.toml
[project]
name = "gekiatsubot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord-buttons-plugin==0.1.2",
    "gspread==6.1.2",
    "gtts>=2.5.3",
    "numpy>=2.0.2",
    "pillow==10.4.0",
    "py-cord>=2.6.1",
    "pynacl>=1.5.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "opencv-python>=4.10.0.84",
    "mysql-connector-python>=9.0.0",
    "ffmpeg-python>=0.2.0",
]
line-length = 88
indent-width = 3
